requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.0.1
python-dateutil==2.8.2
Werkzeug==2.3.7
email-validator==2.1.0